Refactors BFormFieldAsync widget structure
Improves the structure of the BFormFieldAsync widget by removing the BoringRxWatcher wrapper around the entire widget and moving it inside the BFutureBuilder.

This change allows for more efficient updates and rendering of the form field, especially when dealing with asynchronous computations. It also simplifies the widget's logic and makes it easier to maintain.